refactor(backend-service-game): move Writable
