+++ packages/backend/services/backend-service-game/src/foundation/common/application/models/Writable.ts
